chore: reorganize code into a workspace (#24)
Moving all code into a workspace for `core` and `rust`. The goal is to allow for better interop for future languages by having a `core` implementation and language specific bindings.

Refactor some of the existing code so that macros and utility functions are created in the modules they are used. Updated for that leveraged the function `lit` to leverage traits.